third_party/blink/web_tests/external/wpt/pointerevents/pointerevent_pointer_boundary_events_after_removing_last_over_element-expected.txt

This is a testharness.js-based test.
[FAIL] After removing the root element in the shadow under the cursor, pointerleave events should be targeted outside the shadow, but redundant pointerenter events should not be fired
  assert_equals: expected "pointerover@div#shadowHost, pointerout@div#shadowHost, pointerleave@div#shadowHost, pointerleave@div#containerOfShadowHost, pointerover@body, pointermove@body" but got "pointerover@div#shadowHost, pointerenter@body, pointerenter@div#containerOfShadowHost, pointerenter@div#shadowHost, pointerover@div#shadowHost, pointerenter@body, pointerenter@div#containerOfShadowHost, pointerenter@div#shadowHost, pointerover@body, pointerenter@body, pointermove@body"
Harness: the test ran to completion.
